# target/classes/api.yml
openapi: 3.0.1
info :
  title: "twitter"
  version : "1.0"
  contact:
      name : "Souvik"
      email: "deysouvik2014@gmail.com"
servers:
  - url: '//localhost:8083/api/v1.0/tweets'
paths:
...
  /users/all:
    get:
      tags:
        - search
      description: Get all users
      operationId: getAllUser
      parameters:
        - $ref: '#/components/parameters/tokenDetail'
        - $ref: '#/components/parameters/page'
      responses:
        200:
            description: 'all user list'
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/userList'
        500:
            description: 'internal server error'
        400:
            description: 'error details'
            content:
              application/json:
                  schema:
                    $ref: '#/components/schemas/error'
...
components:
  parameters:
    page:
      name: page
      description: no of content to show
      schema:
        type: integer
        format: int32
      in: header
...
    tokenDetail:
      name: Token-key
      in: header
      schema:
        type: string
...
  schemas:
...
    userList:
      title: user List
      type: array
      items: 
        $ref: '#/components/schemas/user'
...
    error:
      title: Error
      type: string
...
    user:
      title: user
      type: object
      properties:
        UserId:
          type: string
        First Name:
            type: string
        last name:
            type: string
        Email:
          type: string
        ContactNumber:
          type: integer
          format: int32